# Conflicts:
#	src/main/java/Test.java
#	src/main/java/entity/Order.java

+++ src/main/java/entity/Order.java
@@ -3,6 +3,7 @@ package entity;
 import enum_types.OrderStatus;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
@@ -37,12 +38,12 @@ public class Order {
     private Collection<Item> items = new ArrayList<>();
 
     @Column(name = "SUMMARY")
-    private Integer summary;
+    private BigDecimal summary;
 
     public Order() {
     }
 
-    public Order(String manager, String buyer, Date date, OrderStatus status, Integer summary) {
+    public Order(String manager, String buyer, Date date, OrderStatus status, BigDecimal summary) {
         this.manager = manager;
         this.buyer = buyer;
         this.date = date;
@@ -98,12 +99,7 @@ public class Order {
         this.items = items;
     }
 
-    public Integer getSummary() {
-        return summary;
-    }
+    public BigDecimal getSummary() {return summary;}
 
-    public void setSummary(Integer summary) {
-        this.summary = summary;
-    }
-
+    public void setSummary(BigDecimal summary) {this.summary = summary;}
 }
